Create Regression code
This is analysis I undertook for the City of Austin to understand if there is a relationship between code violations and the City's home repair program at the Census tract level. Ideally, there would be a strong relationship as the home repair program is designed to help residents with code violations, but because the 2 programs haven't been promoted together, there is a moderate to weak relationship.